backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
pandas
numpy
openai
python-dotenv
